Gambling module fixes, removed gambling service
