+++ SoraBot-v2/SoraBot-v2/Module/GamblingModule.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 using Discord.Commands;
+using Discord.Rest;
 using SoraBot_v2.Data;
 using SoraBot_v2.Services;
 
@@ -9,47 +10,77 @@ namespace SoraBot_v2.Module
     [Name("Gambling")]
     public class GamblingModule : ModuleBase<SocketCommandContext>
     {
+        private CoinService _coinService;
+        private DiscordRestClient _discordRestClient;
+
+        public GamblingModule(CoinService coinService, DiscordRestClient restClient)
+        {
+            _coinService = coinService;
+            _discordRestClient = restClient;
+        }
+
         [Command("coinflip", RunMode = RunMode.Async), Alias("cf"),
          Summary("Flips a coin!")]
         public async Task FlipCoin(int bet, [Remainder]string side)
         {
-            var soraContext = new SoraContext();
-            var restClient = new Discord.Rest.DiscordRestClient();
-            var coinService = new CoinService(restClient);
-            var gamblingService = new GamblingService();
-            var userDb = Utility.GetOrCreateUser(Context.User.Id, soraContext);
-
-            if (bet > coinService.GetAmount(Context.User.Id))
+            using (var soraContext = new SoraContext())
             {
-                await ReplyAsync($"You only have {userDb.Money} Coins.");
-                return;
-            }
+                var userDb = Utility.GetOrCreateUser(Context.User.Id, soraContext);
 
-            if (string.Equals(side, "heads", StringComparison.CurrentCultureIgnoreCase)
-                || string.Equals(side, "tails", StringComparison.CurrentCultureIgnoreCase))
-            {
-                int score = gamblingService.GetResult(side);
-                if (score == 1)
+                if (bet > _coinService.GetAmount(Context.User.Id))
                 {
-                    //won
-                    await Context.Channel.SendMessageAsync("You won!");
-                    userDb.Money += bet * 2;
-                    await soraContext.SaveChangesAsync();
+                    await ReplyAsync($"You only have {userDb.Money} Coins.");
+                    return;
+                }
+
+                if (string.Equals(side, "heads", StringComparison.CurrentCultureIgnoreCase)
+                    || string.Equals(side, "tails", StringComparison.CurrentCultureIgnoreCase))
+                {
+                    int score = GetResult(side);
+                    if (score == 1)
+                    {
+                        //won
+                        await Context.Channel.SendMessageAsync("You won!");
+                        userDb.Money += bet * 2;
+                        await soraContext.SaveChangesAsync();
+                    }
+                    else
+                    {
+                        //lost
+                        await Context.Channel.SendMessageAsync("You lost!");
+                        userDb.Money -= bet;
+                        await soraContext.SaveChangesAsync();
+                    }
                 }
                 else
                 {
-                    //lost
-                    await Context.Channel.SendMessageAsync("You lost!");
-                    userDb.Money -= bet;
-                    await soraContext.SaveChangesAsync();
+                    await Context.Channel.SendMessageAsync($"{side} is not a valid option.");
                 }
             }
-            else
+        }
+
+        public int GetResult(string side)
+        {
+            int score = 0;
+            var random = new Random();
+            bool result = random.Next(100) % 2 == 0;
+            switch (result)
             {
-                await Context.Channel.SendMessageAsync($"{side} is not a valid option.");
+                case true:
+                    if (string.Equals(side, "heads", StringComparison.CurrentCultureIgnoreCase))
+                    {
+                        score = 1;
+                    }
+                    break;
+
+                case false:
+                    if (string.Equals(side, "tails", StringComparison.CurrentCultureIgnoreCase))
+                    {
+                        score = 1;
+                    }
+                    break;
             }
-
-
+            return score;
         }
     }
 }
